Assignment implementation.

+++ course-02/project/image-filter-starter-code/src/util/util.ts
@@ -1,5 +1,8 @@
 import fs from 'fs';
 import Jimp = require('jimp');
+import { resolve } from 'bluebird';
+
+const urlExist = require("url-exist");
 
 // filterImageFromURL
 // helper function to download, filter, and save the filtered image locally
@@ -10,15 +13,21 @@ import Jimp = require('jimp');
 //    an absolute path to a filtered image locally saved file
 export async function filterImageFromURL(inputURL: string): Promise<string>{
     return new Promise( async resolve => {
-        const photo = await Jimp.read(inputURL);
-        const outpath = '/tmp/filtered.'+Math.floor(Math.random() * 2000)+'.jpg';
-        await photo
-        .resize(256, 256) // resize
-        .quality(60) // set JPEG quality
-        .greyscale() // set greyscale
-        .write(__dirname+outpath, (img)=>{
-            resolve(__dirname+outpath);
-        });
+        try {
+            const photo = await Jimp.read(inputURL);
+            const outpath = '/tmp/filtered.'+Math.floor(Math.random() * 2000)+'.jpg';
+            await photo
+            .resize(256, 256) // resize
+            .quality(60) // set JPEG quality
+            .greyscale() // set greyscale
+            .write(__dirname+outpath, (img)=>{
+                resolve(__dirname+outpath);
+            });
+        }
+        catch (error) {
+            console.error(`error during filterImageFromURL: ${error}`);
+            resolve(null);
+        }
     });
 }
 
@@ -29,6 +38,19 @@ export async function filterImageFromURL(inputURL: string): Promise<string>{
 //    files: Array<string> an array of absolute paths to files
 export async function deleteLocalFiles(files:Array<string>){
     for( let file of files) {
-        fs.unlinkSync(file);
+        if(fs.existsSync(file)) {
+            fs.unlinkSync(file);
+        }
     }
+}
+
+export async function isURLValid(image_url: String) : Promise<boolean> {
+
+    var res = await urlExist(image_url);
+    if ( ! res ) {
+        console.error(`urlExist(${image_url}) failed`);
+        return false;
+    }
+
+    return res;
 }
